Converted everything to ES Module

+++ events/emoji.js
@@ -1,28 +1,28 @@
-module.exports = {
+export default {
 
     eventType: 'messageCreate',
     callback: async (msg) => {
 
-        if (msg.author.bot) return;
+        if (msg.webhookId || msg.author.bot) return;
+        if (!(msg.content.length >= 3 && msg.content.startsWith(';') && msg.content.endsWith(';'))) return;
 
-        if (msg.content.length >= 3 && msg.content.startsWith(';') && msg.content.endsWith(';')) {
+        const emojiName = msg.content.replaceAll(';', '');
 
-            let emojiName = msg.content.replaceAll(';', '');
+        for (const [id, emoji] of msg.guild.emojis.cache.entries()) {
+            if (emoji.name === emojiName) {
 
-            for (const [id, emoji] of msg.guild.emojis.cache.entries()) {
-                if (emoji.name === emojiName) {
-                    
-                    const webhook = await msg.channel.createWebhook(msg.member.displayName, {
-                        avatar: msg.author.displayAvatarURL({size: 4096, dynamic: true}),
-                    });
+                if (!msg.guild.me.permissions.has('MANAGE_WEBHOOKS')) {
+                    return await msg.reply('Please enable "Manage Webhooks" permission for me in order to use this feature!');
+                };
 
-                    await webhook.send(emoji.toString());
-                    await webhook.delete();
-                    await msg.delete();
-                    return;
-                }
-            }
-            
-        }
+                const webhook = await msg.channel.createWebhook(msg.member.displayName, {
+                    avatar: msg.author.displayAvatarURL({size: 4096, dynamic: true}),
+                });
+                await webhook.send(emoji.toString());
+                // await webhook.delete();
+                await msg.delete();
+                return;
+            };
+        };
     }
 };
